djangoapp/requirements.txt
Django>=5.0.6,<5.1
psycopg2-binary>=2.9.9,<2.10
camelot-py>=0.11.0,<0.12
PyPDF2>=3.0.1,<3.1
numpy==1.26.4
opencv-python>=4.9.0.80,<4.9.1
ghostscript>=0.7,<0.8
celery>=5.4.0,<5.5
celery-progress>=0.3,<0.4
redis>=5.0.4,<5.1
gunicorn>=22.0.0,<22.1